+++ spec/ManagerSpec.php
@@ -1,10 +1,8 @@
 <?php
 namespace spec\BeatSwitch\Lock;
 
-use BeatSwitch\Lock\ActionAlias;
 use BeatSwitch\Lock\Callers\SimpleCaller;
 use BeatSwitch\Lock\Drivers\Driver;
-use BeatSwitch\Lock\Roles\SimpleRole;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
 
@@ -27,9 +25,19 @@ class ManagerSpec extends ObjectBehavior
 
     function it_can_instantiate_a_new_lock_instance_for_a_role()
     {
-        $this->setRole('editor');
-
         $this->getRoleLock('editor')->shouldBeAnInstanceOf('BeatSwitch\Lock\Roles\RoleLock');
+    }
+
+    function it_can_make_a_caller_lock_aware()
+    {
+        $this->makeCallerLockAware(new SimpleCaller('users', 1))
+            ->shouldBeAnInstanceOf('BeatSwitch\Lock\Callers\Caller');
+    }
+
+    function it_can_make_a_role_lock_aware()
+    {
+        $this->makeRoleLockAware('guest')
+            ->shouldBeAnInstanceOf('BeatSwitch\Lock\Roles\Role');
     }
 
     function it_can_set_action_aliases()